refactor into util methods
